+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changes
 0.5.0 (unreleased)
 ==================
 
+- Connections have a new ``read_only_cursor`` method for obtaining a
+  `database cursor <http://initd.org/psycopg/docs/cursor.html>`_ for
+  selecting data and for using the `cursor's mogrify method
+  <http://initd.org/psycopg/docs/cursor.html#cursor.mogrify>`_.
+
 - The helpers for setting up full-text search indexes now accept a
   config argument to specify the name of a PostgreSQL full-text search
   configuration.