+++ CHANGES.rst
@@ -4,9 +4,27 @@ Changes
 0.5.0 (unreleased)
 ==================
 
+- The ``newt.db.search`` module has a new ``read_only_cursor``
+  function for obtaining a `database cursor
+  <http://initd.org/psycopg/docs/cursor.html>`_ for selecting data and
+  for using the `cursor's mogrify method
+  <http://initd.org/psycopg/docs/cursor.html#cursor.mogrify>`_.
+
 - The helpers for setting up full-text search indexes now accept a
   config argument to specify the name of a PostgreSQL full-text search
   configuration.
+
+- The batch search methods (``search_batch``, and ``where_batch``) now
+  allow the arguments parameter to be omitted, which is useful when
+  substitutions have been be applied with a `cursor mogrify method
+  <http://initd.org/psycopg/docs/cursor.html#cursor.mogrify>`_ before
+  calling them.
+
+- The search methods (``search``, ``search_batch``, ``where`` and
+  ``where_batch``) now accept binary queries.  This is allow
+  substitutions to be applied with a `cursor mogrify method
+  <http://initd.org/psycopg/docs/cursor.html#cursor.mogrify>`_ before
+  calling them.
 
 
 0.4.0 (2017-03-25)